Added the functionality to allow the director to edit the his queue's names and categories from the edit page.
On line 136, I added a name to the input and called it "queueNameEdited" to be able to use it inside the views. I also assigned the value of this input to the queue name of the queue.

On line 158, I added a name to the input and called it "categoriesEdited" to be able to use it inside the views. I also assigned the value of this input to the categories of the queue.