Significant modification to the guest_view_uuid function (see description)
The old function retrieved the data based on the queue uuid.
The current function retrieves the data based on the directors uuid and then retrieves all the queues related to this director and all their categories and sends the categories to the form class to generate the categories_list.
The function creates a new guest object when the user submits